testcases for entry syncing and balance check

+++ factom_cli_tests/loadnodes.py
@@ -1,18 +1,24 @@
 
 import os
 import time
+import unittest
 
 from nose.plugins.attrib import attr
 
 from cli_objects.factom_cli_create import FactomCliCreate
 from cli_objects.factom_chain_objects import FactomChainObjects
 
+
 from helpers.helpers import create_random_string, read_data_from_json
 from random import randint
 
-class LoadNodes():
+class LoadNodes(unittest.TestCase):
     data = read_data_from_json('shared_test_data.json')
     addresses = read_data_from_json('addresses.json')
+    factomd_address = addresses['factomd_address']
+    factomd_address_custom_list = [addresses['factomd_address_0'],addresses['factomd_address_1'], addresses['factomd_address_2'], addresses['factomd_address_3'],addresses['factomd_address_4'],
+                                   addresses['factomd_address_5'],addresses['factomd_address_6']]
+
 
     def __init__(self):
         self.factom_cli_create = FactomCliCreate()
@@ -25,36 +31,52 @@ class LoadNodes():
         self.ecrate = self.factom_cli_create.get_factom_change_entry_credit_conversion_rate()
         self.entry_creds_wallet1 = self.factom_cli_create.import_address_from_factoid(
             self.data['ec_wallet_address'])
-        self.entry_creds_wallet2 = "EC3LYKyMyoqVNezfPjgqBamDfwSZsETYYtGs9YYvwdwbMwWnnSrd"
-
-
-
+        self.entry_creds_wallet2 = self.factom_cli_create.create_entry_credit_address()
 
     def make_chain_and_check_balance(self):
-        for i in xrange(200):
+        chain_flags_list = ['-C ']
+        for i in xrange(10):
             path = os.path.join(os.path.dirname(__file__), '../test_data/testfile')
-            self.factom_cli_create.buy_ec(self.first_address, self.entry_creds_wallet2, '1000')
-
-
-            name_1 = create_random_string(5)
-            name_2 = create_random_string(5)
-            for i in range(200):
+            self.factom_cli_create.buy_ec(self.first_address, self.entry_creds_wallet2, '100')
+            time.sleep(10)
+            for i in range(10):
                 with open('output_file', 'wb') as fout:
                     fout.write(os.urandom(randint(100, 5000)))
                     path = fout.name
-                chain_id = self.factom_chain_object.force_make_chain_from_binary_file_and_receive_chain_id(self.entry_creds_wallet2, path, name_1, name_2)
-
-                for i in range(200):
+                name_1 = create_random_string(5)
+                name_2 = create_random_string(5)
+                names_list = ['-n', name_1, '-n', name_2]
+                chain_id = self.factom_chain_object.make_chain_from_binary(self.entry_creds_wallet2, path, names_list,flag_list=chain_flags_list)
+                for i in range(10):
                     with open('output_file', 'wb') as fout:
                         fout.write(os.urandom(randint(100, 5000)))
                         path = fout.name
                     name_1 = create_random_string(5)
                     name_2 = create_random_string(5)
-                    self.factom_chain_object.add_entries_to_chain(self.entry_creds_wallet2, path, chain_id, name_1, name_2)
-                    os.remove(path)
+                    names_list = ['-e', name_1, '-e', name_2]
+                    self.factom_chain_object.add_entries_to_chain(self.entry_creds_wallet2, path, chain_id, names_list)
+                    if os._exists(path):
+                        os.remove(path)
                     time.sleep(10)
-            time.sleep(5)
-        time.sleep(10)
+                time.sleep(5)
+            time.sleep(10)
+            self.balance_on_nodes([self.entry_creds_wallet2])
+            self.transactions_on_nodes([self.entry_creds_wallet2])
 
 
+    def balance_on_nodes(self,address_list):
+        for address in address_list:
+            balance_1 = self.factom_cli_create.check_wallet_address_balance(address)
+            for factomd_address_custom in self.factomd_address_custom_list:
+                self.factom_cli_create.change_factomd_address(factomd_address_custom)
+                balance_2 = self.factom_cli_create.check_wallet_address_balance(address)
+                self.assertTrue(balance_1 == balance_2, "mismatch in balance. Balance of default server=%s, but server %s=%s, " % (balance_1, factomd_address_custom,balance_2))
 
+
+    def transactions_on_nodes(self,address_list):
+        for address in address_list:
+            transactions_1 = self.factom_cli_create.list_transactions_by_address(address)
+            for factomd_address_custom in self.factomd_address_custom_list:
+                self.factom_cli_create.change_factomd_address(factomd_address_custom)
+                transactions_2 = self.factom_cli_create.list_transactions_by_address(address)
+                self.assertTrue(transactions_1 == transactions_2, "mismatch in transactions. Transaction of default server=%s, %s=%s" % (transactions_1, factomd_address_custom, transactions_2))
